Make it work on Chromium.
It didn't work on Chromium browsers previously.
When a peer joins a room and sends isPeerReady message, the existing peers sends an offer, but the ready peer doesn't receive it.
I logged the offer and found out that it was empty (only had the toUserId).
The issue was in spreading the session description and adding the toUserId fied to it, so I removed that and added toUser to sendMessage as a separate parameter, and after testing it on Chromium browsers, it worked.